Split Donations from DonationsOrClaims
DonationOrClaim represents a tigh-coupling of the Donation and Claim
entities in the codebase. While there exist associations between the
two in the schema, and in the form of shareable code, coupling them
together necessitates logic to infer one or the other based on
conditional logic which can easily be avoided by separating the two
constructs at the namespace level.

This commit formally splits a Donation from DonationsOrClaims.