Split DonorDashboardScreen from DashboardScreen
This commit introduces a DonorDashboardScreen by essentially copying
the DashboardScreenModule and submodules. The rationale behind the
decision is that the needs of a client-dashboard will be different from
that of a donor-dashboard. Not splitting the two will lead to tighter
coupling and less easily negotiable code as time progresses.

While the donor will access the DonorDashboardScreen, the client will
continue accessing the DashboardScreen until a ClientDashboard can be
created.